+++ src/cucu/environment.py
@@ -2,14 +2,15 @@ import datetime
 import hashlib
 import json
 import os
-import shutil
 import sys
 import time
+import traceback
 from functools import partial
 
 from cucu import config, init_scenario_hook_variables, logger
 from cucu.config import CONFIG
 from cucu.page_checks import init_page_checks
+from cucu.utils import ellipsize_filename, take_screenshot
 
 CONFIG.define(
     "FEATURE_RESULTS_DIR",
@@ -31,30 +32,6 @@ CONFIG.define(
 init_page_checks()
 
 
-def generate_image_filename(step_index, step_name):
-    """
-    generate .png image file name that meets these criteria:
-     - hides secrets
-     - escaped
-     - filename does not exceed 255 chars (OS limitation)
-     - uniqueness comes from step number
-    """
-    max_filename = 255 - len(".png")
-    escaped_step_name = CONFIG.hide_secrets(step_name).replace("/", "_")
-    filename = f"{step_index} - {escaped_step_name}"
-
-    if len(filename) > max_filename:
-        ellipsis = "..."
-        # save the last chars as the ending often important
-        end_count = 100
-        front_count = max_filename - (len(ellipsis) + end_count)
-        filename = (
-            filename[:front_count] + ellipsis + filename[-1 * end_count :]
-        )
-
-    return f"{filename}.png"
-
-
 def check_browser_initialized(ctx):
     """
     check browser session initialized otherwise throw an exception indicating
@@ -69,6 +46,8 @@ def before_all(ctx):
     CONFIG["__CUCU_CTX"] = ctx
     CONFIG.snapshot()
     ctx.check_browser_initialized = partial(check_browser_initialized, ctx)
+    for hook in CONFIG["__CUCU_BEFORE_ALL_HOOKS"]:
+        hook(ctx)
 
 
 def after_all(ctx):
@@ -80,7 +59,9 @@ def after_all(ctx):
 def before_feature(ctx, feature):
     if config.CONFIG["CUCU_RESULTS_DIR"] is not None:
         results_dir = config.CONFIG["CUCU_RESULTS_DIR"]
-        ctx.feature_dir = os.path.join(results_dir, feature.name)
+        ctx.feature_dir = os.path.join(
+            results_dir, ellipsize_filename(feature.name)
+        )
         CONFIG["FEATURE_RESULTS_DIR"] = ctx.feature_dir
 
 
@@ -109,7 +90,9 @@ def before_scenario(ctx, scenario):
     ctx.step_timers = {}
 
     if config.CONFIG["CUCU_RESULTS_DIR"] is not None:
-        ctx.scenario_dir = os.path.join(ctx.feature_dir, scenario.name)
+        ctx.scenario_dir = os.path.join(
+            ctx.feature_dir, ellipsize_filename(scenario.name)
+        )
         CONFIG["SCENARIO_RESULTS_DIR"] = ctx.scenario_dir
         os.makedirs(ctx.scenario_dir, exist_ok=True)
 
@@ -146,10 +129,51 @@ def before_scenario(ctx, scenario):
         hook(ctx)
 
 
+def run_after_scenario_hook(ctx, scenario, hook):
+    try:
+        hook(ctx)
+        logger.debug(f"HOOK {hook.__name__}: passed ✅")
+    except Exception as e:
+        # For any after scenario hooks,'hook_failed' status will be 'False' but will attach the error message to scenario.
+        error_message = (
+            f"HOOK-ERROR in {hook.__name__}: {e.__class__.__name__}: {e}\n"
+        )
+        error_message += traceback.format_exc()
+        logger.error(error_message)
+
+
 def after_scenario(ctx, scenario):
     for timer_name in ctx.step_timers:
         logger.warn(f'timer "{timer_name}" was never stopped/recorded')
 
+    run_after_scenario_hook(ctx, scenario, download_mht_data)
+
+    # run after all scenario hooks in 'lifo' order.
+    for hook in CONFIG["__CUCU_AFTER_SCENARIO_HOOKS"][::-1]:
+        run_after_scenario_hook(ctx, scenario, hook)
+
+    # run after this scenario hooks in 'lifo' order.
+    for hook in CONFIG["__CUCU_AFTER_THIS_SCENARIO_HOOKS"][::-1]:
+        run_after_scenario_hook(ctx, scenario, hook)
+
+    CONFIG["__CUCU_AFTER_THIS_SCENARIO_HOOKS"] = []
+
+    if CONFIG.true("CUCU_KEEP_BROWSER_ALIVE"):
+        logger.debug("keeping browser alive between sessions")
+    else:
+        if len(ctx.browsers) != 0:
+            logger.debug("quitting browser between sessions")
+
+        run_after_scenario_hook(ctx, scenario, download_browser_logs)
+
+    cucu_config_filepath = os.path.join(
+        ctx.scenario_logs_dir, "cucu.config.yaml.txt"
+    )
+    with open(cucu_config_filepath, "w") as config_file:
+        config_file.write(CONFIG.to_yaml_without_secrets())
+
+
+def download_mht_data(ctx):
     if not ctx.browsers:
         logger.debug("No browsers - skipping MHT webpage snapshot")
     elif config.CONFIG["CUCU_BROWSER"].lower() != "chrome":
@@ -166,45 +190,26 @@ def after_scenario(ctx, scenario):
             logger.debug(f"Saving MHT webpage snapshot: {mht_filename}")
             browser.download_mht(mht_pathname)
 
-    # run after all scenario hooks in 'lifo' order.
-    for hook in CONFIG["__CUCU_AFTER_SCENARIO_HOOKS"][::-1]:
-        hook(ctx)
 
-    # run after this scenario hooks in 'lifo' order.
-    for hook in CONFIG["__CUCU_AFTER_THIS_SCENARIO_HOOKS"][::-1]:
-        hook(ctx)
+def download_browser_logs(ctx):
+    # close the browser unless someone has set the keep browser alive
+    # environment variable which allows tests to reuse the same browser
+    # session
 
-    CONFIG["__CUCU_AFTER_THIS_SCENARIO_HOOKS"] = []
+    for browser in ctx.browsers:
+        # save the browser logs to the current scenarios results directory
+        browser_log_filepath = os.path.join(
+            ctx.scenario_logs_dir, "browser_console.log.txt"
+        )
 
-    if CONFIG.true("CUCU_KEEP_BROWSER_ALIVE"):
-        logger.debug("keeping browser alive between sessions")
-    else:
-        if len(ctx.browsers) != 0:
-            logger.debug("quitting browser between sessions")
+        os.makedirs(os.path.dirname(browser_log_filepath), exist_ok=True)
+        with open(browser_log_filepath, "w") as output:
+            for log in browser.get_log():
+                output.write(f"{json.dumps(log)}\n")
 
-        # close the browser unless someone has set the keep browser alive
-        # environment variable which allows tests to reuse the same browser
-        # session
-        for browser in ctx.browsers:
-            # save the browser logs to the current scenarios results directory
-            browser_log_filepath = os.path.join(
-                ctx.scenario_logs_dir, "browser_console.log.txt"
-            )
+        browser.quit()
 
-            os.makedirs(os.path.dirname(browser_log_filepath), exist_ok=True)
-            with open(browser_log_filepath, "w") as output:
-                for log in browser.get_log():
-                    output.write(f"{json.dumps(log)}\n")
-
-            browser.quit()
-
-        ctx.browsers = []
-
-    cucu_config_filepath = os.path.join(
-        ctx.scenario_logs_dir, "cucu.config.yaml.txt"
-    )
-    with open(cucu_config_filepath, "w") as config_file:
-        config_file.write(CONFIG.to_yaml_without_secrets())
+    ctx.browsers = []
 
 
 def before_step(ctx, step):
@@ -217,6 +222,8 @@ def before_step(ctx, step):
     ctx.current_step = step
     ctx.current_step.has_substeps = False
     ctx.start_time = time.monotonic()
+
+    CONFIG["__STEP_SCREENSHOT_COUNT"] = 0
 
     # run before all step hooks
     for hook in CONFIG["__CUCU_BEFORE_STEP_HOOKS"]:
@@ -240,40 +247,13 @@ def after_step(ctx, step):
     # and this step has no substeps as in the reporting the substeps that
     # may actually do something on the browser take their own screenshots
     if ctx.browser is not None and ctx.current_step.has_substeps is False:
-        filepath = os.path.join(
-            ctx.scenario_dir, generate_image_filename(ctx.step_index, step.name)
-        )
-
-        # If we've marked an element as the one we're interacting with,
-        # inject a border to highlight that element
-        if (
-            not CONFIG["CUCU_INJECT_ELEMENT_BORDER"]
-            or not CONFIG["__PERTINENT_ELEMENT"]
-        ):
-            ctx.browser.screenshot(filepath)
-        else:
-            border_style = "solid magenta 4px"
-            border_radius = "4px"
-            highlighter = (
-                f'arguments[0].style["border"] = "{border_style}";'
-                f'arguments[0].style["border-radius"] = "{border_radius}";'
-            )
-            ctx.browser.execute(highlighter, CONFIG["__PERTINENT_ELEMENT"])
-            ctx.browser.screenshot(filepath)
-            clear_highlight = (
-                'arguments[0].style["border"] = "";'
-                'arguments[0].style["border-radius"] = "";'
-            )
-            ctx.browser.execute(clear_highlight, CONFIG["__PERTINENT_ELEMENT"])
-            CONFIG["__PERTINENT_ELEMENT"] = None
-
-        if CONFIG["CUCU_MONITOR_PNG"] is not None:
-            shutil.copyfile(filepath, CONFIG["CUCU_MONITOR_PNG"])
+        take_screenshot(ctx, step.name, label=f"After {step.name}")
 
     # if the step has substeps from using `run_steps` then we already moved
     # the step index in the run_steps method and shouldn't do it here
     if not step.has_substeps:
         ctx.step_index += 1
+        CONFIG["__STEP_SCREENSHOT_COUNT"] = 0
 
     if CONFIG.bool("CUCU_IPDB_ON_FAILURE") and step.status == "failed":
         ctx._runner.stop_capture()