+++ src/cucu/formatter/junit.py
@@ -8,8 +8,10 @@ import bs4
 from behave.formatter.base import Formatter
 from behave.model_core import Status
 from bs4.formatter import XMLFormatter
+from tenacity import RetryError
 
 from cucu.config import CONFIG
+from cucu.utils import ellipsize_filename
 
 
 class CucuJUnitFormatter(Formatter):
@@ -37,6 +39,7 @@ class CucuJUnitFormatter(Formatter):
         date_now = datetime.now()
         self.feature_results = {
             "name": escape(feature.name),
+            "foldername": escape(ellipsize_filename(feature.name)),
             "tests": 0,
             "errors": 0,
             "failures": 0,
@@ -57,10 +60,13 @@ class CucuJUnitFormatter(Formatter):
             self.current_scenario_results["time"] = str(
                 round(self.current_scenario_duration, 3)
             )
+            hook_failed = self.current_scenario.hook_failed
+            if hook_failed:
+                status = "errored"
+            else:
+                status = self.current_scenario.compute_status().name
 
-            status = self.current_scenario.compute_status().name
             self.current_scenario_results["status"] = status
-
             failures = []
 
             if status == "failed" and self.current_scenario_traceback:
@@ -74,8 +80,20 @@ class CucuJUnitFormatter(Formatter):
                     )
 
                 failures += [
-                    f"{self.current_step.keyword} {self.current_step.name}"
+                    f"{self.current_step.keyword} {self.current_step.name} (after {round(self.current_step.duration, 3)}s)"
                 ]
+
+                if error := self.current_step.exception:
+                    if isinstance(error, RetryError):
+                        error = error.last_attempt.exception()
+
+                    if len(error.args) > 0 and isinstance(error.args[0], str):
+                        error_class_name = error.__class__.__name__
+                        error_lines = error.args[0].splitlines()
+                        error_lines[0] = f"{error_class_name}: {error_lines[0]}"
+                    else:
+                        error_lines = [repr(error)]
+                    failures += error_lines
 
                 if CONFIG["CUCU_JUNIT_WITH_STACKTRACE"] == "true":
                     failures += traceback.format_tb(
@@ -99,6 +117,9 @@ class CucuJUnitFormatter(Formatter):
             "failure": None,
             "skipped": None,
         }
+        self.current_scenario_results["foldername"] = escape(
+            ellipsize_filename(scenario.name)
+        )
         if scenario.tags:
             self.current_scenario_results["tags"] = ", ".join(scenario.tags)
 
@@ -107,7 +128,7 @@ class CucuJUnitFormatter(Formatter):
             scenario_name
         ] = self.current_scenario_results
 
-        # we write out every new scenario into the JUnit results output which
+        # we write out every new scenario into the JUnit results output
         # which allows us to have a valid JUnit XML results file per feature
         # file currently running even if the process crashes or is killed so we
         # can still generate valid reports from every run.
@@ -148,6 +169,7 @@ class CucuJUnitFormatter(Formatter):
         given a feature results dictionary that looks like so:
             {
                 "name": "name of the feature",
+                "foldername": "",
                 "tests": 0,
                 "errors": 0,
                 "failures": 0,
@@ -155,6 +177,7 @@ class CucuJUnitFormatter(Formatter):
                 "timestamp": "",
                 "scenarios": {
                     "scenario name": {
+                        "foldername": "",
                         "tags": "DOM-3435, testrail(3366,45891)",
                         "status": "passed/failed/skipped",
                         "time": "0.0000":
@@ -173,6 +196,7 @@ class CucuJUnitFormatter(Formatter):
                 ordered = [
                     "classname",
                     "name",
+                    "foldername",
                     "tests",
                     "errors",
                     "failures",
@@ -190,6 +214,7 @@ class CucuJUnitFormatter(Formatter):
         soup = bs4.BeautifulSoup()
         testsuite = bs4.Tag(name="testsuite")
         testsuite["name"] = results["name"]
+        testsuite["foldername"] = results["foldername"]
         testsuite["timestamp"] = results["timestamp"]
 
         junit_dir = CONFIG["CUCU_JUNIT_DIR"]
@@ -239,6 +264,7 @@ class CucuJUnitFormatter(Formatter):
             testcase = bs4.Tag(name="testcase")
             testcase["classname"] = results["name"]
             testcase["name"] = scenario_name
+            testcase["foldername"] = scenario["foldername"]
             if "tags" in scenario:
                 testcase["tags"] = scenario["tags"]
             testcase["status"] = scenario["status"]