+++ src/cucu/reporter/html.py
@@ -14,7 +14,7 @@ import jinja2
 from cucu import format_gherkin_table, logger
 from cucu.ansi_parser import parse_log_to_html
 from cucu.config import CONFIG
-from cucu.environment import generate_image_filename
+from cucu.utils import ellipsize_filename, get_step_image_dir
 
 
 def escape(data):
@@ -100,6 +100,7 @@ def generate(results, basepath, only_failures=False):
     CONFIG.snapshot()
     reported_features = []
     for feature in features:
+        feature["folder_name"] = ellipsize_filename(feature["name"])
         if show_status:
             print("F", end="", flush=True)
         scenarios = []
@@ -115,6 +116,7 @@ def generate(results, basepath, only_failures=False):
         total_scenarios_passed = 0
         total_scenarios_failed = 0
         total_scenarios_skipped = 0
+        total_scenarios_errored = 0
         feature_started_at = None
 
         reported_features.append(feature)
@@ -122,8 +124,10 @@ def generate(results, basepath, only_failures=False):
 
         if feature["status"] not in ["skipped", "untested"]:
             # copy each feature directories contents over to the report directory
-            src_feature_filepath = os.path.join(results, feature["name"])
-            dst_feature_filepath = os.path.join(basepath, feature["name"])
+            src_feature_filepath = os.path.join(results, feature["folder_name"])
+            dst_feature_filepath = os.path.join(
+                basepath, feature["folder_name"]
+            )
             shutil.copytree(
                 src_feature_filepath, dst_feature_filepath, dirs_exist_ok=True
             )
@@ -131,8 +135,11 @@ def generate(results, basepath, only_failures=False):
         for scenario in scenarios:
             CONFIG.restore()
 
+            scenario["folder_name"] = ellipsize_filename(scenario["name"])
             scenario_filepath = os.path.join(
-                basepath, feature["name"], scenario["name"]
+                basepath,
+                feature["folder_name"],
+                scenario["folder_name"],
             )
 
             scenario_configpath = os.path.join(
@@ -177,6 +184,8 @@ def generate(results, basepath, only_failures=False):
                 total_scenarios_failed += 1
             elif scenario["status"] == "skipped":
                 total_scenarios_skipped += 1
+            elif scenario["status"] == "errored":
+                total_scenarios_errored += 1
 
             step_index = 0
             scenario_started_at = None
@@ -184,16 +193,38 @@ def generate(results, basepath, only_failures=False):
                 if show_status:
                     print("s", end="", flush=True)
                 total_steps += 1
-                image_filename = generate_image_filename(
-                    step_index, step["name"]
-                )
-                image_filepath = os.path.join(scenario_filepath, image_filename)
+                image_dir = get_step_image_dir(step_index, step["name"])
+                image_dirpath = os.path.join(scenario_filepath, image_dir)
 
                 if step["name"].startswith("#"):
                     step["heading_level"] = "h4"
 
-                if os.path.exists(image_filepath):
-                    step["image"] = urllib.parse.quote(image_filename)
+                if os.path.exists(image_dirpath):
+                    _, _, image_names = next(os.walk(image_dirpath))
+                    images = []
+                    for image_name in image_names:
+                        words = image_name.split("-", 1)
+                        index = words[0].strip()
+                        try:
+                            # Images with label should have a name in the form:
+                            # 0000 - This is the image label.png
+                            label, _ = os.path.splitext(words[1].strip())
+                        except IndexError:
+                            # Images with no label should instead look like:
+                            # 0000.png
+                            # so we default to the step name in this case.
+                            label = step["name"]
+
+                        images.append(
+                            {
+                                "src": urllib.parse.quote(
+                                    os.path.join(image_dir, image_name)
+                                ),
+                                "index": index,
+                                "label": label,
+                            }
+                        )
+                    step["images"] = sorted(images, key=lambda x: x["index"])
 
                 if "result" in step:
                     if step["result"]["status"] in ["failed", "passed"]:
@@ -205,11 +236,10 @@ def generate(results, basepath, only_failures=False):
                         if scenario_started_at is None:
                             scenario_started_at = timestamp
                             scenario["started_at"] = timestamp
-                        step["result"][
-                            "time_offset"
-                        ] = datetime.utcfromtimestamp(
+                        time_offset = datetime.utcfromtimestamp(
                             (timestamp - scenario_started_at).total_seconds()
                         )
+                        step["result"]["time_offset"] = time_offset
 
                     scenario_duration += step["result"]["duration"]
 
@@ -308,12 +338,14 @@ def generate(results, basepath, only_failures=False):
         feature["total_scenarios_passed"] = total_scenarios_passed
         feature["total_scenarios_failed"] = total_scenarios_failed
         feature["total_scenarios_skipped"] = total_scenarios_skipped
+        feature["total_scenarios_errored"] = total_scenarios_errored
 
     keys = [
         "total_scenarios",
         "total_scenarios_passed",
         "total_scenarios_failed",
         "total_scenarios_skipped",
+        "total_scenarios_errored",
         "duration",
     ]
     grand_totals = {}
@@ -367,7 +399,7 @@ def generate(results, basepath, only_failures=False):
     feature_template = templates.get_template("feature.html")
 
     for feature in reported_features:
-        feature_basepath = os.path.join(basepath, feature["name"])
+        feature_basepath = os.path.join(basepath, feature["folder_name"])
         os.makedirs(feature_basepath, exist_ok=True)
 
         scenarios = []
@@ -392,7 +424,9 @@ def generate(results, basepath, only_failures=False):
 
         for scenario in scenarios:
             steps = scenario["steps"]
-            scenario_basepath = os.path.join(feature_basepath, scenario["name"])
+            scenario_basepath = os.path.join(
+                feature_basepath, scenario["folder_name"]
+            )
             os.makedirs(scenario_basepath, exist_ok=True)
 
             scenario_output_filepath = os.path.join(