+++ src/cucu/steps/text_steps.py
@@ -1,7 +1,7 @@
 from cucu import fuzzy, helpers, step
 from cucu.browser.frames import try_in_frames_until_success
 from cucu.steps import step_utils
-from cucu.utils import text_in_current_frame
+from cucu.utils import take_saw_element_screenshot, text_in_current_frame
 
 
 def find_text(ctx, name, index=0):
@@ -17,13 +17,17 @@ def find_text(ctx, name, index=0):
         the WebElement that matches the provided arguments or None if none found
     """
     ctx.check_browser_initialized()
-    return fuzzy.find(
+    element = fuzzy.find(
         ctx.browser,
         name,
         ["*"],
         index=index,
         direction=fuzzy.Direction.LEFT_TO_RIGHT,
     )
+
+    take_saw_element_screenshot(ctx, "text", name, index, element)
+
+    return element
 
 
 # Also update the line number in the  scenario: `User gets the right stacktrace for steps using step helpers` when changing the code below.